% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSP.R
\name{aesDT}
\alias{aesDT}
\title{Helper function to plot ggplot directly from data.table.}
\usage{
aesDT(...)
}
\arguments{
\item{...}{Arguments to be passed (explicitly) to ggplot aesthetic}
}
\description{
See \url{from http://r.789695.n4.nabble.com/ggplot2-and-data-table-td3743483.html}
  for details
  
Credit to Timothée Carayol
}

