% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mtraceR.R
\name{BF}
\alias{BF}
\title{Calculate Bayes Factor}
\usage{
BF(lmLs = NULL)
}
\arguments{
\item{lmLs}{Numeric vector with log marginal likelihood values of the models}
}
\value{
A data.frame with log Bayes factors (LBF), the models' ranks
(mod.rank) and the models' probability (mod.prob).
}
\description{
\code{BF} calculates log Bayes factors (LBF) of all models compared against
the model that has highest log likelihood. The models are then ranked based
on LBF (mod.rank) and the model probability (mod.prob) is calculated.
}
\examples{
# From migrate-n tutotial: 'Comparison of gene flow models using
# Bayes Factors' (Beerli 2010)
 mod.comp <- BF(c(-4862.85, -4860.58, -4863.08, -4887.25))
 mod.comp
}

