% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSP.R
\name{meta.BSP}
\alias{meta.BSP}
\title{Sum parameter values from different parameters}
\usage{
meta.BSP(data = NULL, dir.out = NULL, params = NULL, locus = "max",
  save2disk = TRUE)
}
\arguments{
\item{data}{The data generated by \code{BSP}}

\item{dir.out}{The local path to store the results. If NULL (default) then
\code{dir.out=wd()}}

\item{params}{A vector of parameter numbers to be combined and plotted}

\item{locus}{The locus to be considered}

\item{save2disk}{Whether to save results to disk (default: TRUE)}
}
\value{
A list with the combined data (as data.table) and a BSP.
}
\description{
\code{meta.BSP} takes as input data equal to the first element of the output
  from \code{BSP} and combines (i.e. sum) the values of the parameters passed 
  with the argument \code{params}. This could be useful, for example, when 
  the analysis includes populations from a meta-population system and the user
  wants to calculate the size of meta-population (i.e. the total).
  
When multiple loci are used in the analysis, the argument \code{locus} can 
  be used to indicate which locus should be used. If \code{locus="max"},
  then the lastgest locus number is used (which corresponds to the sum over 
  all loci for multi-locus analyses).
}

