% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtraceR.R
\name{ms_3popsSt}
\alias{ms_3popsSt}
\title{Example of \code{mtrace} output}
\format{a \code{list} with 3 named elelments.}
\source{
Pacioni et al. In prep.
}
\usage{
data(ms_3popsSt)
}
\description{
\code{mtrace} output from a run (with 7 replicates) using 
  microsatellite data with 4 sampling sites grouped in three populations 
  connected by a stepping stone migration model.
}
