% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mtraceR.R
\name{mtDNA_4popsSt}
\alias{mtDNA_4popsSt}
\title{Example of \code{mtrace} output}
\format{a \code{list} with 3 named elelments.}
\source{
Pacioni et al. In prep.
}
\usage{
data(mtDNA_4popsSt)
}
\description{
\code{mtrace} output from a run (with no replicate) using sequence
  data (mtDNA) with 4 sampling sites connected by a stepping stone migration
  model between either sites 1 or 2 and 3 and 4
  (i.e. migration model: {* * 0 *   * * 0 *   0 0 * *   * * * *}).
}

