% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mtraceR.R
\name{mtrace}
\alias{mtrace}
\title{Diagnostics of migrate-n MCMC chains}
\usage{
mtrace(heating = TRUE, nchain = 4, burn.in = 0.1, npop = 1,
  estim.npar = NULL, trim = TRUE, bayesallfile = NULL, dir.in = NULL,
  dir.out = NULL, save2disk = TRUE)
}
\arguments{
\item{heating}{Whether heating is used: (\code{TRUE} or \code{FALSE})}

\item{nchain}{Number of chains if \code{heating=TRUE}}

\item{burn.in}{Length of burn-in as proportion of MCMC (default: 0.1)}

\item{npop}{The number of populations in the analysis}

\item{estim.npar}{The number of estimated parameters (when a custom migration
matrix is used)}

\item{trim}{Whether trim data to estimated parameters (default:  TRUE)}

\item{bayesallfile}{The name of the bayesallfile (default: NULL)}

\item{dir.in}{The local folder containing migrate-n output files (default: NULL)}

\item{dir.out}{The local path to store the results. If NULL (default) then
\code{dir.out=dir.in}}

\item{save2disk}{Whether to save results to disk (default: TRUE)}
}
\value{
a list with the ESS for each locus, a summary of ESS across all loci
  and Gelamn's diagnostic if more replicates were run. See details for
  additional outputs when \code{save2dis=TRUE}.
}
\description{
\code{mtrace} parses migrate-n output file (bayesallfile) and calculates the
Effective Sample Size (ESS) for each locus. If \code{save2dis=TRUE} these
are also saved to disk together with a PDF with the trace and density plots
for each locus and each replicate.
}
\details{
If there is more than 1 replicate, \code{mtrace} returns also Gelman and
Rubin's convergence diagnostic for each locus. When \code{save2dis=TRUE}
relevant diagnostic plots are also saved as a separate PDF.

When no bayesallfile name and directory are provided (default), the output
file is selected with an iteractive window. Alternatively, when the directory
path is provided with \code{dir.in} and \code{bayesallfile=NULL}, the latter
is automatically set to "bayesallfile.gz". If the name of the file is provided,
but \code{dir.in=NULL}, the directory path is selected iteractively.

If \code{dir.out=NULL} (default) then \code{dir.out=dir.in}.
}

