\name{mtraceR-package}
\alias{mtraceR-package}
\alias{mtraceR}
\docType{package}
\title{
\packageTitle{mtraceR}
}
\description{
\packageDescription{mtraceR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mtraceR}
\packageIndices{mtraceR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mtraceR}

Maintainer: \packageMaintainer{mtraceR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
