% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mtraceR.R
\docType{package}
\name{mtraceR}
\alias{mtraceR}
\alias{mtraceR-package}
\title{mtraceR: an R package to analyse migration-n output}
\description{
\code{mtraceR} can be used to visualise trace of MCMC chains, calculate
Effective Sample Size (ESS), Gelman and Rubin's convergence diagnostic (and
plots) or Bayes factor, and generate skyline plots.
}
\details{
The main function \code{mtrace} is essentially a wrapper for the R package
\code{coda} optimised for migrate-n outputs. It is intended to facilitate and
speed up post-analysis data processing.

COMPLETE DESCRIPTION/DETAILS
}
\section{Documentations}{

Use \code{help(package = "mtraceR")} for a list of \code{mtraceR} functions
and their specific documentations.

A more detailed description of the package and functions can be opened with:
\code{vignette(package="mtraceR", topic="mtraceR-package")}.

More vignettes may be come available in the future. Use
\code{vignette(package="mtraceR")} to see all the available vignettes.
}

\section{Citation}{

If you use \code{mtraceR}, please cite:
...
}

\section{Get in touch}{

Please, use \url{https://github.org/carlopacioni/mtraceR/issues} to report
any issues with \code{mtraceR}. If unsure, or for feedback, contact us at:
carlo.pacioni 'at' gmail.com or anders???.
}

