% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSP.R
\name{multi.BSP}
\alias{multi.BSP}
\title{Generate multiple Bayesian skyline plots within one figure}
\usage{
multi.BSP(data1 = NULL, data2 = NULL, col1 = "black", col2 = "red",
  y.title = "", dir.out = NULL, save2disk = TRUE)
}
\arguments{
\item{data1, data2}{The data generated by \code{BSP} to be used for the BSPs}

\item{col1, col2}{The colour to be used for the BSPs}

\item{y.title}{The title for the y-axis (default="")}

\item{dir.out}{The local path to store the results. If NULL (default) then
\code{dir.out=wd()}}

\item{save2disk}{Whether to save results to disk (default: TRUE)}
}
\value{
A list with the combined data (as data.table) and a BSP.
}
\description{
\code{multi.BSP} takes as 2 separate input data equal to the first element of 
  the output from \code{BSP} reduced to a unique parameter and a unique locus 
  and generates a figure where two BSP are concurrently plotted.
}
