% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMigrate.R
\name{runMigrate}
\alias{runMigrate}
\title{Run migrate-n from within R}
\usage{
runMigrate(
  runOptions = "mpirun --use-hwthread-cpus",
  migratePath = "",
  migrateExecutable = "migrate-n-mpi",
  nProcessors = "auto",
  pathParmfile = "./",
  parmfile = "parmfile"
)
}
\arguments{
\item{runOptions}{Character vector to pre-append to the command being issued. 
The default assume that you are running migrate in parallel using openMPI
 (`"mpirun --use-hwthread-cpus"`)}

\item{migratePath}{Character vector with the path to migrate excutable}

\item{migrateExecutable}{The name of the executable (default `migrate-n-mpi`)}

\item{nProcessors}{The number of processor to be used. If `"auto"`, the maximum 
number of processors available will be used}

\item{pathParmfile}{The path to the param file (default `"./"`}

\item{parmfile}{The name of the parmfile (default `"parmfile"`)}
}
\description{
This is a convenience function to run migrate-n from within R
}
