% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BSP.R
\name{BSP}
\alias{BSP}
\title{Generate Bayesian skyline plots}
\usage{
BSP(dir.in = NULL, skylinefile = NULL, dir.out = NULL, all.loci = TRUE,
  overall = TRUE, params = 1, gen = 1, mu = 1, save2disk = TRUE)
}
\arguments{
\item{dir.in}{The local folder containing skylinefile files (default: NULL)}

\item{skylinefile}{The name of the skylinefile (default: NULL)}

\item{dir.out}{The local path to store the results. If NULL (default) then
\code{dir.out=dir.in}}

\item{all.loci}{Whether the parameter should be plotted for each locus}

\item{overall}{Whether the parameter should be plotted for the sum over all
loci (when \code{nloci>1})}

\item{params}{A vector of parameter numbers to be plotted}

\item{gen}{The generation time in unit of time (e.g. years)}

\item{mu}{Mutation rate expressed in the same unit as \code{gen}}

\item{save2disk}{Whether to save results to disk (default: TRUE)}
}
\value{
A list where the first element is a data.table with skyline data and
  the others are the BSPs.
}
\description{
\code{BSP} generates Bayesian skyline plots for each analysed locus (when
  relevant) and the sum over all loci. Plots are returned as ggplot objects
  and, when \code{save2disk==TRUE}, pdf files.
}
\details{
\code{BSP} will also return the data used to generate the BSPs as a data.table
  object and, when \code{save2disk==TRUE}, a csv file. These data are a 'clean
  up' version of these contained in the skyline file along with the upper and
  lower limits. The column "Time" is the rescaled "Age" (from migrate output)
  when \code{gen} and \code{mu} are provided. The column "Ne" is the
  effective population size, calculated as theta/(x * mu).
}

